通信用データ型の定義とひな形の作成
