+++ frontend/src/home_and_profile/CreatePost.tsx
@@ -1,0 +1,10 @@
+export default function CreatePost() {
+    return (
+        <div className="createPost">
+            <form>
+                <textarea />
+                <button>投稿する</button>
+            </form>
+        </div>
+    )
+}
