+++ frontend/src/home_and_profile/Post.tsx
@@ -1,10 +1,11 @@
 import './Post.css'
+import { postData } from "../../../common/connectionData";
 
-export default function Post() {
+export default function Post( data : postData ) {
     return (
         <div className="post">
             <p>Title<br/>
-            Body</p>
+            {data.content} </p>
         </div>
     )
 }
